add tests for the add_dir and remove_dir function

+++ tests/data/test_interface.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 import pytest
+
 from pplkit.data.interface import DataInterface
 
 tmpdir = Path(__file__).parents[1] / "tmp"
@@ -17,12 +18,11 @@ def data():
 @pytest.fixture(scope="class", autouse=True)
 def rm_tmpdir_after_tests():
     yield
-    shutil.rmtree(tmpdir)
+    if tmpdir.exists():
+        shutil.rmtree(tmpdir)
 
 
-@pytest.mark.parametrize(
-    "fextn", [".json", ".yaml", ".pkl", ".csv", ".parquet"]
-)
+@pytest.mark.parametrize("fextn", [".json", ".yaml", ".pkl", ".csv", ".parquet"])
 def test_data_interface(data, fextn):
     dataif = DataInterface(tmp=tmpdir)
     if fextn in [".csv", ".parquet"]:
@@ -32,3 +32,20 @@ def test_data_interface(data, fextn):
 
     for key in ["a", "b"]:
         assert np.allclose(data[key], loaded_data[key])
+
+
+def test_add_dir():
+    dataif = DataInterface()
+    assert len(dataif.keys) == 0
+    dataif.add_dir("tmp", tmpdir)
+    assert len(dataif.keys) == 1
+    assert hasattr(dataif, "tmp")
+    assert hasattr(dataif, "load_tmp")
+    assert hasattr(dataif, "dump_tmp")
+
+
+def test_remove_dir():
+    dataif = DataInterface(tmp=tmpdir)
+    assert len(dataif.keys) == 1
+    dataif.remove_dir("tmp")
+    assert len(dataif.keys) == 0
